g_adk/requirement.txt
google-adk
langchain
langchain-community
sentence-transformers
pymilvus
uvicorn
